encode checked tx
