+++ libs/cosmos-sdk/x/auth/types/stdtx.go
@@ -451,8 +451,6 @@ func EthereumTxDecode(b []byte, tx interface{}) error {
 	return rlp.DecodeBytes(b, tx)
 }
 
-
-
 // MarshalYAML returns the YAML representation of the signature.
 func (ss StdSignature) MarshalYAML() (interface{}, error) {
 	var (
