+++ x/evm/types/utils.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"encoding/binary"
 	"fmt"
-	authtypes "github.com/okex/exchain/libs/cosmos-sdk/x/auth/types"
 	"math/big"
 	"strings"
 
@@ -15,9 +14,6 @@ import (
 	ethcrypto "github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/rlp"
 	"github.com/okex/exchain/app/crypto/ethsecp256k1"
-	"github.com/okex/exchain/libs/cosmos-sdk/codec"
-	sdk "github.com/okex/exchain/libs/cosmos-sdk/types"
-	sdkerrors "github.com/okex/exchain/libs/cosmos-sdk/types/errors"
 	"github.com/pkg/errors"
 	"golang.org/x/crypto/sha3"
 )
@@ -525,71 +521,6 @@ func DecodeResultData(in []byte) (ResultData, error) {
 		return ResultData{}, err
 	}
 	return data, nil
-}
-
-// ----------------------------------------------------------------------------
-// Auxiliary
-
-func TxDecoder(cdc *codec.Codec) sdk.TxDecoder {
-	return func(txBytes []byte) (sdk.Tx, error) {
-		var err error
-		if len(txBytes) == 0 {
-			return nil, sdkerrors.Wrap(sdkerrors.ErrTxDecode, "tx bytes are empty")
-		}
-		payloadDecoder := payloadTxDecoder(cdc)
-
-		//----------------------------------------------
-		//----------------------------------------------
-		// 1. try sdk.CheckedTx
-		var chkTx sdk.Tx
-		if chkTx, err = authtypes.DecodeCheckedTx(txBytes, payloadDecoder); err == nil {
-			return chkTx, nil
-		} else {
-			fmt.Printf("DecodeCheckedTx failed:%p %s\n", txBytes, err)
-		}
-
-		return payloadDecoder(txBytes)
-	}
-}
-
-func payloadTxDecoder(cdc *codec.Codec) sdk.TxDecoder {
-	return func(txBytes []byte) (sdk.Tx, error) {
-		var err error
-		if len(txBytes) == 0 {
-			return nil, sdkerrors.Wrap(sdkerrors.ErrTxDecode, "tx bytes are empty")
-		}
-
-		//----------------------------------------------
-		//----------------------------------------------
-		// 2. Try to decode as MsgEthereumTx through RLP
-		var ethTx MsgEthereumTx
-		if err = authtypes.EthereumTxDecode(txBytes, &ethTx); err == nil {
-			return ethTx, nil
-		} else {
-			fmt.Printf("EthereumTxDecode failed: %p %s\n", txBytes, err)
-		}
-
-		//----------------------------------------------
-		//----------------------------------------------
-		// 3. try other concrete message types registered by MakeTxCodec
-		// TODO: switch to UnmarshalBinaryBare on SDK v0.40.0
-		var tx sdk.Tx
-		if v, err := cdc.UnmarshalBinaryLengthPrefixedWithRegisteredUbmarshaller(txBytes, &tx); err == nil {
-			return v.(sdk.Tx), nil
-		} else {
-			fmt.Printf("UnmarshalBinaryLengthPrefixedWithRegisteredUbmarshaller failed:%p %s\n", txBytes, err)
-		}
-
-		//----------------------------------------------
-		//----------------------------------------------
-		// 4. try others
-		if err = cdc.UnmarshalBinaryLengthPrefixed(txBytes, &tx); err == nil {
-			return tx, nil
-		} else {
-			fmt.Printf("UnmarshalBinaryLengthPrefixed failed: %p %s\n", txBytes, err)
-		}
-		return nil, sdkerrors.Wrap(sdkerrors.ErrTxDecode, err.Error())
-	}
 }
 
 // recoverEthSig recovers a signature according to the Ethereum specification and
